removed warning.
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@2856 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0

+++ fem/storage/subarray.hh
@@ -166,7 +166,7 @@ namespace Dune
       typedef CheckVectorInterface< BaseVectorType > __CheckBaseVectorType__;
       typedef CheckIndexMapperInterface< IndexMapperType > __CheckIndexMapperType__;
 
-      assert( baseVector_.size() == indexMapper_.range() );
+      assert( baseVector_.size() == (int) indexMapper_.range() );
     }
